requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==3.5.0
decorator==4.4.2
diffusers==0.33.1
dill==0.3.8
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2024.12.0
gTTS==2.5.4
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.7
pooch==1.8.2
proglog==0.1.11
propcache==0.3.1
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydub==0.25.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.0.0
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
